+++ core/pioreactor/background_jobs/led_automation.py
@@ -59,14 +59,15 @@ class LEDAutomationJob(AutomationJob):
     ) -> None:
         super(LEDAutomationJob, self).__init__(unit, experiment)
 
-        self.add_to_published_settings(
-            "duration",
-            {
-                "datatype": "float",
-                "settable": True,
-                "unit": "min",
-            },
-        )
+        if "duration" not in self.published_settings:
+            self.add_to_published_settings(
+                "duration",
+                {
+                    "datatype": "float",
+                    "settable": True,
+                    "unit": "min",
+                },
+            )
 
         self.skip_first_run = skip_first_run
         self.edited_channels: set[pt.LedChannel] = set()