+++ frontend/src/Leader.jsx
@@ -227,7 +227,7 @@ function DirectoryNavigatorCard({leaderHostname}) {
     <Card sx={{minHeight: "300px"}}>
       <CardContent sx={{p:2}}>
         <Typography variant="h6" component="h2">
-          <Box fontWeight="fontWeightRegular">File browser</Box>
+          <Box fontWeight="fontWeightRegular">System file browser</Box>
         </Typography>
         <Box sx={{maxHeight: "450px", overflowY: "scroll"}}>
           <PathViewer path=
@@ -301,7 +301,7 @@ function DirectoryNavigatorCard({leaderHostname}) {
 
           sx={{textTransform: "none"}}
         >
-          <DownloadIcon fontSize="small" sx={textIcon} /> Export
+          <DownloadIcon fontSize="small" sx={textIcon} /> Export system files
         </Button>
       </CardActions>
     </Card>
@@ -413,7 +413,7 @@ function LeaderCard({leaderHostname}) {
                 fontWeight: 500,
               }}
               gutterBottom>
-              <PioreactorIconWithModel badgeContent={"🔷"}  sx={{verticalAlign: "middle", mr: "3px", mb: "2px"}} />
+              <PioreactorIconWithModel badgeContent={"◎"}  sx={{verticalAlign: "middle", mr: "3px", mb: "2px"}} />
               {unit}
             </Typography>
             <Tooltip title={indicatorLabel} placement="right">